Version 2.6.6.008
Colby now includes version.php which holds the version number. This approach gives the version number an obvious location while not forcing a change to Colby.php each time the version number changes.

This checkin also adds the ColbyVersion MySQL function to return the version when the database was last updated so we can detect whether database updates are needed after Colby is updated.