Version 2.7.0.003
Implemented methods in ColbyArchiver and added the createFileWithFileId method.

Since the next step is implementing the blog but that may take a while I added a page to colby to run unit tests: /colby/test/. So far this only tests that the ColbyArchiver class works and doesn't test for error handling or any other functionality. But after this, adding tests just means adding them to the ajax handler that runs the tests.