Version 2.7.1.010
Archiving now uses hashes to make sure that changes are only made when the previous version of the file is the version the caller thinks it should be. If this is not the case, the file was probably changed in another browser window or by another user and file should be updated and merged before saving. One way to do this is to open the file in yet another browser, merge your changes in, and then save from thre.