Version 2.9.0
Overview: So much work has been done that it's irresponsible not to push this to the server. After more work and finalization, this will become version 3.0.0 and then converting other apps will commence. For now, although things work fine, and that's why this is being committed, this is still a sort of intermediate milestone.