Version 2.9.0.002
Overview:

While working with the system it became apparent that serializing named classes (not stdClass) to data files is very risky. Any class renames would break every existing archive. If serialized to data files, class names can never be renamed again. Furthermore, there were other technical reasons that having the serialized data directly on core objects was difficult. For one, it was prohibitively difficult to add pxrivate data that was not supposed to be serialized.

So with this checkin the ColbyPageModel class hosts the data internally but does not own it. The data is stored in a stdClass that is shared with other classes each of which can modify the data in ways they are supposed to. This stdClass is then the object that is serialized to the data file and since it is a stdClass there's extremely low risk of it ever changing names.

In addition to these changes, the ColbyPageModel class now provides setters and getters for all its properties. This was a good idea because all of the users of this data were writing duplicate and error prone code to get and set data. Not only will this prevent errors in the data, but now errors when getting and setting properties will cause exceptions and be caught sooner.