Version 3.0.0.003
Overview:

This checkin add support for Javascript time formatting. This is a nice feature because Javascript is aware of the users location and timezone and can easily give them time data locally. The server just uses UTC timestamps (times 1000, Javascript uses milliseconds) which Javascript can easily process and transfer to dates.

The time support is really good as it constantly updates the times and counts up the seconds for a time less that 60 seconds in the past, then minutes up to an hour, then other nice formats for times up to two days in the past.

To use it, simple have any element with the "time" class and a "data-timestamp" property:

    <span class="time" data-timestamp="1355732450000"></span>

Once the page is loaded the content of the span will be changed to a human readable time and updated as necessary.

modified:   handlers/handle,developer,test.php
-   Added unit tests to test the new functionality to make sure it keeps working.

modified:   javascript/Colby.js
-   Added the new time functionality to the Colby Javascript class.