Version 1
Overview
--------

This checkin is a change to the way version numbers are handled. This checkin
represents version 1, and each new version will increment this number. The
version numbers will get quite high and that is fine.

This is being done for ease of use in comparing version numbers for to enable
upgrade code to run when necessary. It's also happening because before this
version numbers were changed in a haphazard fashion with no real logic behind
changed the major, middle, or minor version number except for what "felt right."

All checkins should be done on a working or development branch and the version
number used for those checkins should be in the following format, assuming the
most recent master branch version number is 5:

5.001
5.002
5.003
...

When the work on the working branch is ready to be checked into the master
branch then the version number will be incremented to 6 on the working branch
and after that is merged into the master branch so the master branch version
number becomes 6. The master branch version number will be incremented by one
every time the master branch changes, no matter how many intermediate working
branch checkins were made.

Deployment Strategy
-------------------

To make future updates easier and avoid detecting if whether the
'ColbySchemaVersionNumber' database function or the 'COLBY_VERSION_NUMBER'
PHP constant is available, there will be no (or very few) changes made to
Colby until this version is deployed on all sites.

The only changes will be the most minimal changes required for deployment on
current sites.

File changes
------------

modified:   handlers/handle,developer,configuration,ajax,setup-database.php
-   Changed the 'ColbyVersion' function to 'ColbySchemaVersionNumber' which
    now returns the Colby version number for the last time the database schema
    was updated. This may differ from the most recent Colby version number if
    the database hasn't been updated in a while, which is intentional to make
    sure all of the database updates are still applied if the admin forgets to
    do database updates between version of Colby.

modified:   version.php
-   Changed 'COLBY_VERSION' to 'COLBY_VERSION_NUMBER' and set the value to 1.