Version 53.21
This checkin involves a relatively simple but destructively diff confusing change of renaming all of the current "header" and "footer" sections to "page header" and "page footer". Don't bother looking at what Git thinks has happened because it is very confused.

Suffice it so say that it's important to use the term "page header" and "page footer" in these cases because there will be various "article header" and "article footer" sections in the future which may be named something more like "blog post header" and "blog post footer".