Add the `CBDB::optional` and `CBDB::stringToSQL` functions
This checkin is small but is very much a part of the "functional
programming" migration. The `CBDB::stringToSQL` function was added
to replace the `ColbyConvert::textToSQL` function which does not
return a value wrapped in single quotes and therefore cannot be used
to generate comma separated lists of values embeddable in a "WHERE IN"
clause.

The `CBDB::optional` function gives any other function the ability to
convert `null` values into the SQL 'NULL' value.

These two functions together with array_map provide powerful ways of
generating lists of values to be used in SQL.