CBTextBoxView: Add default "None" theme
This theme really is intended to represent "no theme" but it actually
has a healthy set of styles to make it look reasonably good but
still not "themed". It uses the default fonts and colors, but is styled
to still look presentable with reasonable padding, line spacing, and
margins. The goal is not to theme "no theme" but to make sure that it
doesn't actively look bad through neglect.