Remove URLsForCSS from CBTextBoxViewTheme
These were removed because a view had to be re-saved to notice that
the CSS URLs had changed. This meant changing the CSS URLs would not
affect previously saved views. All changes to a theme should immediately
affect all current views. CSS files can be included in the styles
using the @import directive. Since it's used almost exclusively for
fonts, this is fine.