Allow duplicate URIs and remove typeID and groupID (big change)
This checkin represents an update first to remove the deprecated typeID
and groupID fields from the ColbyPages table.

Also, this checkin removes the unique index on the ColbyPages URI column.
This was done because all of the work done to ensure unique URIs in the
database was causing performance issues. Not on every day stuff, but on
rule of a million stuff. It inserted more queries and would even siply stop
changes from being allowed with not much one could do about it.

Now, all URLs are allows. The oldest published page with a given URL will
be the page provided for the URL. Duplicate published pages with the same
URL will pointed out in the admin area. Then a user can decide what to do
about it.