Change request handling (big change)
The change is both minor and major. It's minor in that the request handling
rules have not really changed. It's major in that the code has changed
drastically and a lot of bugs have been fixed.

Canonical URLs are known and redirected to more often. The case of
URIs in the database is respected and redirected to. The organization
of the code is vastly better.