Add CBThemeEditorFactory JavaScript class
We are about to have many different themes all of which behave the exact
same way. I thought about having them all share a theme class and giving
the themes a "kind" but to do this correctly would mean changing the
models table which is a big deal. Instead they can share this new theme
editing widget and then if some themes need customization they aren't
block from having it.

The CBThemedTextViewTheme was updated to use the new shared theme editor.