Responsive: Update CBImageLinkView
Image was aligning left when not inside another view. There was a purity
to this behavior but it was never really useful. So now the image is centered
within the greater element bounds which will reach to 100% as it is just
a div without width set.

It's better behavior in all scenarios and may bring a useful future life
to this view.