Update CBContainerView
At some point in the near future all computers are going to have retina
displays. It is not worth it to go our of our way to make our code more
complex to support non-retina displays.

Furthermore, JPEG images that are retina can look just as good and be
the same size as non-retina images at much lower quality. Often they
actually look better, especially if they contain text.

After much investigation this change modifies the container view
to only show retina images. The change was made in a very reversible
way but that's unlikely to be needed.