Release version 189
This update deprecates some old system funtions and constants. It adds some
site administration information. A new way of creating administrative widgets
was added.

The ability was added to remotely administer the source code with a new CBGit
class and the update administration page was improved so that when site
update is requested that a database backup is automatically made, the site
will update to the latest source code available for the site, and then
the update script will be run.

These changes make it much faster to deploy site changes.