Release version 232
Optimize menu classes, CBWellKnownThemeForContent

The classes CBMenu, CBMenuItem, CBThemedMenuView are modernized by
removing uncalled functions and fixing function declarations and
comments.

The class CBWellKnownThemeForContent was also modernized in the same
ways and in addition its install() function was improved for clarity.

Modernize model editor

The model editor is used to edit models for preferences, menus, and
themes. The editor was moved to the admin page class method of
displaying its pages. It's dependencies were also updated with the
new URLs used to send the user to a model editor.

Upgrade CBAdminPageForEditingModels

CBAdminPageForEditingModels is changed to use CBUISpecSaver to save
the currently edited model.