Fix width in CBTextView2StandardLayout
First the width was set to 680px. This would cause the view to be too
wide at times. Then the max-width was set to 640px. This would cause
the view to be too narrow at times.

Now the inner div has this rule, flex: 0 1 680px. This means the div
has a default size of 680px, and it can't grow, but it can shrink if
if needs to.