Update CBAdminPageForModelImport
When importing a CSV file, the models are now saved immediately
of by creating a task for each model. Models saves are supposed to be
relatively quick and this is especially true for models that are
likely to be imported in large quantities.

A model class that needs to do more should create tasks in its
CBModels_willSave() function.