Add cache busting support to the .htaccess template
CSS and JavaScript files with names like "Colby.v358.js" will be
redirected to their unversioned URIs. This allows the versioned URIs
to to bust caches.

When developers update CSS and JavaScript files they should update
the version number of the URI in the PHP file.

The Colby class is an example of how to do this.