Release version 402
Pruning of model versions is now performed by a task that is
restarted each time a model is saved. Using a task allows the
process to be slightly more complicated without hindering the saving
process. Specifically the change makes sure that when a model is
pruned at least three versions are kept so that one bad save will
allow a model to be recovered.

a30c974e Add CBModelPruneVersionsTask
17a693df Add CBModelPruneVersionsTaskTests
c494d81c Use CBModelPruneVersionsTask in CBModels
8c9ab33a Fix CBModelPruneVersionsTask for models that currently have no versions