Update CBImageVerificationTask.php
There was a bug in the image resizing code that if a requested size was larger than the original size, the original image would be copied to the requested size file. The bag was that if the file type requested was different than the original we would still do this and change the exception. So a webp file might actually be a jpeg file or an png file.

I added a check of all files with the webp extension to make sure they are actually webp files or delete them to be regenerated upon request.