Release version 675.63
This version added the administrative title to the CB_CBView_Hero1. This title is used to let administrators know what the view is for and is not shown to the public.

CB_CBView_Hero1 now also provides thumbnail images and a description for the editing environment.

2f8082ca Update CB_CBView_Hero1.php
edf1002a Update CB_CBView_Hero1Editor